refactor: _middleware.tsの実装をmiddleware.tsに統合
重み付きA/Bテスト・QAパラメータ・リセット機能を持つ実装に差し替え、_middleware.tsを削除。
